+++ packages/frontend/src/components/council-dashboard/hook/use-council-filter.ts
@@ -6,6 +6,7 @@ import type {
   CouncilFilterFormatter,
 } from '@/components/council-dashboard/type'
 import type { CouncilMeeting } from '@/types/council-meeting'
+import type { MemberType } from '@twreporter/congress-dashboard-shared/lib/constants/council-member'
 // lodash
 import { find } from 'lodash'
 const _ = {
@@ -17,6 +18,8 @@ const useCouncilFilter = (meetings: CouncilMeeting[]) => {
   const [filterValues, setFilterValues] = useState<CouncilFilterModalValueType>(
     {
       meeting: latestMeetingTerm,
+      administrativeDistrict: [],
+      type: 'all',
       constituency: [],
       party: [],
     }
@@ -36,8 +39,22 @@ const useCouncilFilter = (meetings: CouncilMeeting[]) => {
       const constituency = Array.isArray(filterValues.constituency)
         ? filterValues.constituency.map((idString: string) => Number(idString))
         : []
+      const types = typeof filterValues.type === 'string' && filterValues.type !== 'all'
+          ? [filterValues.type as MemberType]
+          : []
+      const administrativeDistricts = Array.isArray(
+        filterValues.administrativeDistrict
+      )
+        ? filterValues.administrativeDistrict
+        : []
 
-      return { meetingId, partyIds, constituency }
+      return {
+        meetingId,
+        partyIds,
+        constituency,
+        types,
+        administrativeDistricts,
+      }
     },
     [meetings]
   )